# apps/messaging-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO Messaging API
  description: API for OGCIO Messaging Service
  version: 0.1.0
components:
  schemas:
    def-0:
      type: object
      properties:
        statusCode:
          type: number
        code:
          type: string
        error:
          type: string
        message:
          type: string
      title: HttpError
paths:
...
  /api/v1/providers/emails/:
    get:
      tags: &a2
        - Providers - Emails
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          format: uuid
                          type: string
                        name:
                          type: string
                        host:
                          type: string
                        port:
                          type: number
                        username:
                          type: string
                        password:
                          type: string
                        throttle:
                          type: number
                        fromAddress:
                          type: string
                      required:
                        - id
                        - name
                        - host
                        - port
                        - username
                        - password
                        - fromAddress
                required:
                  - data
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    post:
      tags: *a2
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                host:
                  type: string
                port:
                  type: number
                username:
                  type: string
                password:
                  type: string
                throttle:
                  type: number
                fromAddress:
                  type: string
              required:
                - name
                - host
                - port
                - username
                - password
                - fromAddress
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        format: uuid
                        type: string
                    required:
                      - id
                required:
                  - data
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/providers/emails/{providerId}:
    get:
      tags: *a2
      parameters:
        - schema:
            type: string
          in: path
          name: providerId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        format: uuid
                        type: string
                      name:
                        type: string
                      host:
                        type: string
                      port:
                        type: number
                      username:
                        type: string
                      password:
                        type: string
                      throttle:
                        type: number
                      fromAddress:
                        type: string
                    required:
                      - id
                      - name
                      - host
                      - port
                      - username
                      - password
                      - fromAddress
                required:
                  - data
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    put:
      tags: *a2
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  format: uuid
                  type: string
                name:
                  type: string
                host:
                  type: string
                port:
                  type: number
                username:
                  type: string
                password:
                  type: string
                throttle:
                  type: number
                fromAddress:
                  type: string
              required:
                - id
                - name
                - host
                - port
                - username
                - password
                - fromAddress
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: providerId
          required: true
      responses:
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    delete:
      tags: *a2
      parameters:
        - schema:
            type: string
          in: path
          name: providerId
          required: true
      responses:
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...